mudando as classes para arquivos separados
