+++ ContaTerminal/entities/Menu.java
@@ -1,0 +1,9 @@
+package entities;
+
+public class Menu {
+    private static String[] menu = {"1 - Cadastrar", "2 - Consultar", "3 - Finalizar"};    
+   
+	public static String[] getMenu(){
+		return menu;
+	}
+}
